Rename to `SectorsPieChart`
