+++ src/components/SectorsPieChart.tsx
